FastAPI/requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
blinker==1.6.2
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.4
debugpy==1.8.0
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.104.0
frozenlist==1.4.0
gitdb==4.0.10
GitPython==3.1.31
gunicorn==21.2.0
h11==0.14.0
httpx==0.25.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.25.2
ipython==8.16.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==8.3.1
jupyter_core==5.3.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib-inline==0.1.6
mdurl==0.1.2
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
openai==0.28.1
packaging==23.1
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==11.0.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pypdf==3.16.4
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.1.1
requests==2.28.2
rich==13.3.4
scikit-learn==1.3.1
scipy==1.11.3
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
stack-data==0.6.3
starlette==0.27.0
streamlit==1.27.1
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
tqdm==4.66.1
traitlets==5.11.2
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
uvicorn==0.23.2
validators==0.20.0
wcwidth==0.2.8
wget==3.2
yarl==1.9.2
zipp==3.15.0
fastapi==0.68.0
python-jose[cryptography]==3.3.0
passlib==1.7.4